Mask off support of events for the different event list
Not all events are supported by every release of the Switchtec
firmware.  To solve this, add a evlist_mask field to the event_regs
table which is a bitmask of all supported event list for a given
event.

From Gen 4, a new event list version in GAS was introduced, any
definition change of event will trigger a version increase. For
Gen 3 which doesn't support this version, we have a list of all
known event lists for all the known releases.

For the unknown Gen 3 releases, all events are masked supported.

Check this mask as part of validating and event_ctl IOCTL and when
clearing all the event bits in the ISR.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>