Mask off support of events for the different event list
Not all events are supported by every release of the Switchtec
firmware. To solve this, since Gen4, a new bit in each event header
is introduced to indicate if an event is supported. For Gen3 which
does not support this, we add a ver_mask field to the event_regs
table which is a bitmask of all supported event list versinos for
a given event.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>